+++ packages/lesswrong/lib/components.ts
@@ -1105,8 +1105,9 @@ importComponent("ReviewQuickPage", () => require('../components/review/ReviewQui
 importComponent("NewLongformReviewForm", () => require('../components/review/NewLongformReviewForm'));
 importComponent("ReviewDashboardButtons", () => require('../components/review/ReviewDashboardButtons'));
 importComponent("ReviewPhaseInformation", () => require('../components/review/ReviewPhaseInformation'));
-importComponent("UserReviewsProgressBar", () => require('../components/review/UserReviewsProgressBar'));
-importComponent("ReviewVotingProgressBar", () => require('../components/review/ReviewVotingProgressBar'));
+importComponent("ReviewProgressNominations", () => require('../components/review/ReviewProgressNominations'));
+importComponent("ReviewProgressReviews", () => require('../components/review/ReviewProgressReviews'));
+importComponent("ReviewProgressVoting", () => require('../components/review/ReviewProgressVoting'));
 importComponent("ReviewVotingPageMenu", () => require('../components/review/ReviewVotingPageMenu'));
 importComponent("ReviewVotingCanvas", () => require('../components/review/ReviewVotingCanvas'));
 importComponent("FrontpageReviewWidget", () => require('../components/review/FrontpageReviewWidget'));
